Move qr logic to a separate file
