+++ src/qr.h
@@ -1,0 +1,8 @@
+#include "tox.h"
+#include "native/image.h"
+
+void qr_setup(const char *id_str,
+              uint8_t **qr_data,
+              int *qr_data_size,
+              NATIVE_IMAGE **qr_image,
+              int *qr_image_size);
